+++ .gitignore
@@ -1,3 +1,4 @@
+.vscode
 **/*.idea/
 **/target
 **.DS_Store
